+++ src/Admin/Controller.php
@@ -80,7 +80,7 @@ class Controller
         }
 
         // add links to documentation
-        $links[] = '<a href="https://www.kokoanalytics.com/kb/">' . esc_html__('Documentation', 'koko-analytics') . '</a>';
+        $links[] = '<a href="https://www.kokoanalytics.com/docs/">' . esc_html__('Documentation', 'koko-analytics') . '</a>';
 
         // add link to Pro version, unless already running it
         if (! \defined('KOKO_ANALYTICS_PRO_VERSION')) {