+++ src/Resources/views/settings/tracking.php
@@ -37,7 +37,7 @@ defined('ABSPATH') or exit;
                 </li>
             </ul>
 
-            <p class="description"><?php echo sprintf(wp_kses(__('For some more information about how each of these methods work, read this article on <a href="%1$s">cookie vs. cookieless tracking</a>.', 'koko-analytics'), ['a' => ['href' => true]]), 'https://www.kokoanalytics.com/kb/cookie-vs-cookieless-tracking-methods'); ?></p>
+            <p class="description"><?php echo sprintf(wp_kses(__('For some more information about how each of these methods work, read this article on <a href="%1$s">cookie vs. cookieless tracking</a>.', 'koko-analytics'), ['a' => ['href' => true]]), 'https://www.kokoanalytics.com/docs/tracking/cookie-vs-cookieless-tracking-methods/'); ?></p>
         </fieldset>
     </div>
     <div class="mb-4">